Enable pulseaudio client mode and add version tag.

+++ ledfx_docker/start.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 set -eu
-echo "Cleaning up old PulseAudio files..."
-rm -rf "$HOME/.config/pulse/"* || echo "Failed to clean Pulseaudio files. Check permissions on mounted folder."
-echo "Starting pulseaudio"
-pulseaudio --daemonize=yes --disallow-module-loading --log-target=stderr --disable-shm=yes --exit-idle-time=-1 --load="module-native-protocol-unix auth-anonymous=1 auth-cookie-enabled=0 socket=$HOME/.config/pulse/pulseaudio.socket" --load="module-always-sink"
+
+if [ -n "$PULSEAUDIOCLIENT" ]; then
+  echo "Running in Pulseaudio client mode. Please ensure default source is set correctly."
+else
+  echo "Running in pulseaudio server mode. Cleaning up old PulseAudio files..."
+  rm -rf "$HOME/.config/pulse/"* || echo "Failed to clean Pulseaudio files. Check permissions on mounted folder."
+  echo "Starting pulseaudio"
+  pulseaudio --daemonize=yes --disallow-module-loading --log-target=stderr --disable-shm=yes --exit-idle-time=-1 --load="module-native-protocol-unix auth-anonymous=1 auth-cookie-enabled=0 socket=$HOME/.config/pulse/pulseaudio.socket" --load="module-always-sink"
+fi
+
 sleep 1
 echo "Starting ledfx"
 exec ledfx -c "$HOME/ledfx-config" "$@"
